ESDEV-3511 Refactor Doctrine::isSelectStatement and rename to Doctrine::doesStatementProduceOutput
The original method did not recognize all SQL statements, that produce output and therefore should be treated as "SELECT" statements.
The method is also able to deal with comments in the SQL statements.
The name of the method was misleading, so the function was renamed to doesStatementProduceOutput
(cherry picked from commit 10b1810)