ESDEV-3822 Fix test, which introduced duplicate values in database
The test setup in Integration/Price/BasketConstruct.php introduced duplicate values to table oxobject2group. This will throw an exception, as there is an UNIQUE index on this table now.
The exception is caught now and in case of being an DUPLICATE KEY exception silently discarded.
(cherry picked from commit d420368)