ESDEV-4003 #454 Fix loosing sessions with captchas in subshops
Bug ID 0006476

isCurrentUrl has again a problem that it does not correctly find the matching sub shop url
because it fails for /core/utils/verificationimg.php.

this will destroy the session of the current user because the shopid from the url will not match with that shopid from the session
and then the sessionid will be regenerated.