ESDEV-3641 Deprecate surplus functions, which were only there in the sake of testing and would break things , if used in the shop.
The functions setDbObject, getDbObject and resetTblDescCache were only used in the tests and should never be called by the shop code. But as the were public they could be called and would probably do some harm. These methods were removed and the testing code was refactored.
 The function escapeString is used partly to escape SQL values, partly to escape SQL identifiers. This method was deprecated and should be removed.
 The function isValidFieldName() was deprecated and should be moved to the database adapter class Doctrine
(cherry picked from commit 4536441)